Fix #1710: Docker build args not working
Ported PR https://github.com/fabric8io/fabric8-maven-plugin/pull/1726/
