+++ jkube-kit/enricher/api/src/main/java/io/jkube/maven/enricher/api/util/KubernetesResourceUtil.java
@@ -451,73 +451,6 @@ public class KubernetesResourceUtil {
         return true;
     }
 
-
-    public static boolean setEnvVar(List<EnvVar> envVarList, String name, String value) {
-        for (EnvVar envVar : envVarList) {
-            String envVarName = envVar.getName();
-            if (Objects.equals(name, envVarName)) {
-                String oldValue = envVar.getValue();
-                if (Objects.equals(value, oldValue)) {
-                    return false;
-                } else {
-                    envVar.setValue(value);
-                    return true;
-                }
-            }
-        }
-        EnvVar env = new EnvVarBuilder().withName(name).withValue(value).build();
-        envVarList.add(env);
-        return true;
-    }
-
-    public static String getEnvVar(List<EnvVar> envVarList, String name, String defaultValue) {
-        String answer = defaultValue;
-        if (envVarList != null) {
-            for (EnvVar envVar : envVarList) {
-                String envVarName = envVar.getName();
-                if (Objects.equals(name, envVarName)) {
-                    String value = envVar.getValue();
-                    if (StringUtils.isNotBlank(value)) {
-                        return value;
-                    }
-                }
-            }
-        }
-        return answer;
-    }
-
-    public static boolean removeEnvVar(List<EnvVar> envVarList, String name) {
-        boolean removed = false;
-        for (Iterator<EnvVar> it = envVarList.iterator(); it.hasNext(); ) {
-            EnvVar envVar = it.next();
-            String envVarName = envVar.getName();
-            if (name.equals(envVarName)) {
-                it.remove();
-                removed = true;
-            }
-        }
-        return removed;
-    }
-
-    /**
-     * Convert a map of env vars to a list of K8s EnvVar objects.
-     * @param envVars the name-value map containing env vars which must not be null
-     * @return list of converted env vars
-     */
-    public static List<EnvVar> convertToEnvVarList(Map<String, String> envVars) {
-        List<EnvVar> envList = new LinkedList<>();
-        for (Map.Entry<String, String> entry : envVars.entrySet()) {
-            String name = entry.getKey();
-            String value = entry.getValue();
-
-            if (name != null) {
-                EnvVar env = new EnvVarBuilder().withName(name).withValue(value).build();
-                envList.add(env);
-            }
-        }
-        return envList;
-    }
-
     public static void validateKubernetesMasterUrl(URL masterUrl) throws MojoExecutionException {
         if (masterUrl == null || StringUtils.isBlank(masterUrl.toString())) {
             throw new MojoExecutionException("Cannot find Kubernetes master URL. Have you started a cluster via `mvn jkube:cluster-start` or connected to a remote cluster via `kubectl`?");
